Explicited int division on mask center calculation
Python 3 needs integer division to be called explicitly or else the result is returned as a floating point value. This resulted in maskout[ctridx] accessed with a floating point value and thus raising an exception in Python 3.
Since the integer division operator is the same, it can be fixed without breaking Python 2 compatibility.